% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_query_functions.r
\name{convert_column_class}
\alias{convert_column_class}
\title{convert df column classes based on prefixes}
\usage{
convert_column_class(df, classes = NULL)
}
\arguments{
\item{df}{"data.frame" with data to convert}

\item{classes}{"character" vector with class names, needs to be the same length as
number of columns}
}
\value{
\code{ret_data} "data.frame" with converted column's classes
}
\description{
convert df column classes based on prefixes
}

