% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_query_functions.r
\name{execute_sql_query}
\alias{execute_sql_query}
\title{Execute sql query}
\usage{
execute_sql_query(SQL, db = .__DEFAULT_ODBC_DB_NAME__.,
  schema = .__DEFAULT_FILE_DB_SCHEMA__.)
}
\arguments{
\item{SQL}{"character" query string}

\item{db}{"character" ODBC alias name of database}

\item{schema}{"character" name of DB schema}
}
\description{
Execute sql query
}

