% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_query.r
\docType{methods}
\name{executeSQLQuery,VirtualSQLQuery,missing-method}
\alias{executeSQLQuery,VirtualSQLQuery,missing-method}
\title{Executes SQL query strings}
\usage{
\S4method{executeSQLQuery}{VirtualSQLQuery,missing}(object)
}
\arguments{
\item{object}{object of class 'VirtualSQLQuery'.}
}
\value{
\code{ret_val} data.frame with result of the query
}
\description{
Executes previously prepared SQL queries and returns result
}

