% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_procedure_call_functions.r
\name{generate_procedure_call_strings}
\alias{generate_procedure_call_strings}
\title{convert key values to sql procedure call}
\usage{
generate_procedure_call_strings(proc_n, args, keys)
}
\arguments{
\item{proc_n}{"character" name of the stored procedure}

\item{args}{"character" vector names of the arguments}

\item{keys}{"data.frame"}
}
\value{
\code{SQL} "character" vector of SQL procedure call
strings generated based on given parameters
}
\description{
convert key values to sql procedure call
}

