% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_insert_handler.r
\docType{methods}
\name{getSQLInsertObject}
\alias{getSQLInsertObject}
\alias{getSQLInsertObject,VirtualSQLInsertHandler-method}
\title{Get sql_insert stored in object}
\usage{
getSQLInsertObject(object)

\S4method{getSQLInsertObject}{VirtualSQLInsertHandler}(object)
}
\arguments{
\item{object}{object of class "VirtualSQLInsertHandler"}
}
\value{
\code{sql_insert} object of class "VirtualSQLQuery"

\code{sql_insert} object of class "VirtualSQLQuery"
}
\description{
Returns sql_insert object of class "VirtualSQLQuery"
}
\section{Methods (by class)}{
\itemize{
\item \code{VirtualSQLInsertHandler}: Get sql_insert stored in object

Returns sql_insert object of class "VirtualSQLQuery"
}}

