% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_query.r
\docType{methods}
\name{getSQLQueryColumnNameMap}
\alias{getSQLQueryColumnNameMap}
\alias{getSQLQueryColumnNameMap,VirtualSQLQuery-method}
\title{Retreive SQL query column name map}
\usage{
getSQLQueryColumnNameMap(object)

\S4method{getSQLQueryColumnNameMap}{VirtualSQLQuery}(object)
}
\arguments{
\item{object}{object of class 'VirtualSQLQuery'.}
}
\value{
\code{column_name_map} hash column name map of the datasource to output names

\code{column_name_map} hash column name map of the datasource to output names
}
\description{
Returns hash with mapping of query column names
to return column names
}
\section{Methods (by class)}{
\itemize{
\item \code{VirtualSQLQuery}: Retreive SQL query column name map

Returns hash with mapping of query column names
to return column names
}}

