% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_query.r
\docType{methods}
\name{getSQLQueryKeyColumnNames}
\alias{getSQLQueryKeyColumnNames}
\alias{getSQLQueryKeyColumnNames,VirtualSQLQuery-method}
\title{Get sql query column names}
\usage{
getSQLQueryKeyColumnNames(object)

\S4method{getSQLQueryKeyColumnNames}{VirtualSQLQuery}(object)
}
\arguments{
\item{object}{object of class 'VirtualSQLQuery'.}
}
\value{
\code{key_cols} character vector of constructed SQL query strings.
}
\description{
Returns key column names that are used to construct sql querries.
}
\section{Methods (by class)}{
\itemize{
\item \code{VirtualSQLQuery}: Get sql query column names

Returns key column names that are used to construct sql querries.
}}

