% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_query.r
\docType{methods}
\name{getSQLQueryKeyValues}
\alias{getSQLQueryKeyValues}
\alias{getSQLQueryKeyValues,VirtualSQLQuery-method}
\title{Get sql query column names}
\usage{
getSQLQueryKeyValues(object)

\S4method{getSQLQueryKeyValues}{VirtualSQLQuery}(object)
}
\arguments{
\item{object}{object of class 'VirtualSQLQuery'.}
}
\value{
\code{key_values} data.frame with columns matching key column names()

\code{key_values} data.frame with columns matching key column names()
}
\description{
Returns key values datatable
with keys that will be matched in datastore query
}
\section{Methods (by class)}{
\itemize{
\item \code{VirtualSQLQuery}: Get sql query column names

Returns key values datatable
with keys that will be matched in datastore query
}}

