% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_query_handler.r
\docType{methods}
\name{getSQLQueryObject}
\alias{getSQLQueryObject}
\alias{getSQLQueryObject,VirtualSQLQueryHandler-method}
\title{Get sql_query stored in object}
\usage{
getSQLQueryObject(object)

\S4method{getSQLQueryObject}{VirtualSQLQueryHandler}(object)
}
\arguments{
\item{object}{object of class "VirtualSQLQueryHandler"}
}
\value{
\code{sql_query} object of class "VirtualSQLQuery"

\code{sql_query} object of class "VirtualSQLQuery"
}
\description{
Returns sql_query object of class "VirtualSQLQuery"
}
\section{Methods (by class)}{
\itemize{
\item \code{VirtualSQLQueryHandler}: Get sql_query stored in object

Returns sql_query object of class "VirtualSQLQuery"
}}

