% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_query_functions.r
\name{has_required_columns}
\alias{has_required_columns}
\title{check if data.frame has required columns}
\usage{
has_required_columns(data, required_colnms)
}
\arguments{
\item{data}{"data.frame" with data to test}

\item{required_colnms}{"character" vector with required column names}
}
\value{
\code{result} "logical" TRUE if tested data.frame has required columns
FALSE otherwise
}
\description{
check if data.frame has required columns
}

