% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_query_functions.r
\name{pass_thru_parser}
\alias{pass_thru_parser}
\title{passes argument to result}
\usage{
pass_thru_parser(x)
}
\arguments{
\item{x}{"ANY" pass through function}
}
\description{
passes argument to result
}

