% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_procedure_call.r
\docType{methods}
\name{prepareSQLQuery,VirtualSQLProcedureCall,data.frame-method}
\alias{prepareSQLQuery,VirtualSQLProcedureCall,data.frame-method}
\title{Prepare SQL query strings}
\usage{
\S4method{prepareSQLQuery}{VirtualSQLProcedureCall,data.frame}(object,
  key_values)
}
\arguments{
\item{object}{object of class 'VirtualSQLProcedureCall'.}

\item{key_values}{"data.frame" with query keys}
}
\value{
\code{object} object of class 'VirtualSQLProcedureCall'.
}
\description{
Parses key_values to vector of SQL query strings
}

