% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_query.r
\docType{methods}
\name{prepareSQLQuery}
\alias{prepareSQLQuery}
\alias{prepareSQLQuery,VirtualSQLQuery,data.frame-method}
\title{Prepare SQL query strings}
\usage{
prepareSQLQuery(object, key_values = NA)

\S4method{prepareSQLQuery}{VirtualSQLQuery,data.frame}(object,
  key_values = NA)
}
\arguments{
\item{object}{object of class 'VirtualSQLQuery'.}

\item{key_values}{"data.frame" with query keys}
}
\value{
\code{object} object of class 'VirtualSQLQuery'.
}
\description{
Parses key_values to vector of SQL query strings
}
\section{Methods (by class)}{
\itemize{
\item \code{object = VirtualSQLQuery,key_values = data.frame}: Prepare SQL query strings

Parses key_values to vector of SQL query strings
}}

