% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_insert_handler.r
\name{setSQLInsertObject}
\alias{setSQLInsertObject}
\title{Set sql_insert object in object slot}
\usage{
setSQLInsertObject(object, sql_insert)
}
\arguments{
\item{object}{object of class "VirtualSQLInsertHandler"}

\item{sql_insert}{object of class "VirtualSQLQuery"}
}
\value{
\code{object} object of class "VirtualSQLInsertHandler"
}
\description{
Public method to set sql_insert slot with "VirtualSQLQuery"
class object to be implemented in derived
}

