% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_query_handler.r
\name{setSQLQueryObject}
\alias{setSQLQueryObject}
\title{Set sql_query object in object slot}
\usage{
setSQLQueryObject(object, sql_query)
}
\arguments{
\item{object}{object of class "VirtualSQLQueryHandler"}

\item{sql_query}{object of class "VirtualSQLQuery"}
}
\value{
\code{object} object of class "VirtualSQLQueryHandler"
}
\description{
Public method to set sql_query slot with "VirtualSQLQuery"
class object to be implemented in derived
}

