% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TE.SQLQuery.r
\docType{package}
\name{TE.SQLQuery}
\alias{TE.SQLQuery}
\alias{TE.SQLQuery-package}
\title{Trading Enhancement interface to SQL Databases}
\description{
Package providing classes and functions for handling queries to
SQL databases
}
\section{TE.SQLQuery classes}{

VirtualSQLQuery
}

