% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_insert_handler.r
\docType{class}
\name{VirtualSQLInsertHandler-class}
\alias{VirtualSQLInsertHandler-class}
\title{Virtual S4 class handling VirtualSQLQuery objects}
\description{
Class that is to be inherited by any objects
that will contain VirtualSQLQuery or derived classes
}
\section{Slots}{

\describe{
\item{\code{sql_insert}}{object of class "VirtualSQLQuery"}
}}

