% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_procedure_call.r
\docType{class}
\name{VirtualSQLProcedureCall-class}
\alias{VirtualSQLProcedureCall-class}
\title{Virtual S4 class handling sql procedures calls.}
\description{
Implements handling of access to data via sql stored procedures calls.
Inherits from "VirtualSQLQuery"
}
\section{Slots}{

\describe{
\item{\code{arguments}}{= "character",}

\item{\code{procedure}}{= "character",}
}}

