% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_query.r
\docType{class}
\name{VirtualSQLQuery-class}
\alias{VirtualSQLQuery-class}
\title{Virtual S4 class sql queries handling.}
\description{
Implements handling of access to data via SQL queries.
}
\section{Slots}{

\describe{
\item{\code{db_name}}{"character" ODBC name of the database}

\item{\code{db_schema}}{"character" schema name}

\item{\code{sql_query}}{"character" sql query string}

\item{\code{key_cols}}{"character" key column names of the query}

\item{\code{key_values}}{"data.frame" key values}

\item{\code{query_parser}}{"function" query parsing function}

\item{\code{results_parser}}{"function" result parsing function}
}}

