% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_query_handler.r
\docType{class}
\name{VirtualSQLQueryHandler-class}
\alias{VirtualSQLQueryHandler-class}
\title{Virtual S4 class handling VirtualSQLQuery objects}
\description{
Class that is to be inherited by any objects
that will contain VirtualSQLQuery or derived classes
}
\section{Slots}{

\describe{
\item{\code{sql_query}}{object of class "VirtualSQLQuery"}
}}

