% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_query.r
\docType{methods}
\name{prepareSQLQuery,VirtualSQLQuery,missing-method}
\alias{prepareSQLQuery,VirtualSQLQuery,missing-method}
\title{Prepare SQL query strings}
\usage{
\S4method{prepareSQLQuery}{VirtualSQLQuery,missing}(object)
}
\arguments{
\item{object}{object of class 'VirtualSQLQuery'.}
}
\value{
\code{object} object of class 'VirtualSQLQuery'.
}
\description{
Parses key_values to vector of SQL query strings
}

